Delete id exception handler

+++ dscommerce/src/main/java/com/oliveiralia/dscommerce/services/ProductService.java
@@ -1,16 +1,21 @@
 package com.oliveiralia.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.oliveiralia.dscommerce.dtos.ProductDto;
 import com.oliveiralia.dscommerce.dtos.mappers.ProductMapper;
 import com.oliveiralia.dscommerce.entities.Product;
 import com.oliveiralia.dscommerce.repositories.ProductRepository;
+import com.oliveiralia.dscommerce.services.exceptions.DatabaseException;
 import com.oliveiralia.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -40,14 +45,27 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDto update(Long id, ProductDto dto) {
-		Product entity = repository.getReferenceById(id);
-		ProductMapper.copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return ProductMapper.fromEntity(entity);
-	}
+		try {
+			Product entity = repository.getReferenceById(id);
+			ProductMapper.copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return ProductMapper.fromEntity(entity);
+		}
+		catch (EntityNotFoundException e) {
+			 throw new ResourceNotFoundException("Recurso não encontrado.");
+		}
+ 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado.");
+		}
+		try {
+			repository.deleteById(id);
+		}catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
+		
 	}
 }
